requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
Pillow
requests
mangum